requirements.txt
Adafruit-Blinka==8.10.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-requests==1.12.11
adafruit-circuitpython-rfm9x==2.2.13
adafruit-circuitpython-typing==1.8.3
Adafruit-PlatformDetect==3.38.0
Adafruit-PureIO==1.1.9
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
board==1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.2.2
Flask-SocketIO==5.3.2
frozenlist==1.3.3
idna==3.4
importlib-metadata==5.2.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.3
numpy==1.24.0
opencv-python==4.6.0.66
Pillow==9.3.0
pyftdi==0.54.0
pyserial==3.5
python-engineio==4.3.4
python-socketio==5.7.2
pyusb==1.2.1
requests==2.28.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
six==1.16.0
socketIO-client==0.7.2
sysv-ipc==1.1.0
typing-extensions==4.4.0
urllib3==1.26.13
websocket-client==1.4.2
Werkzeug==2.2.2
yarl==1.8.2
zipp==3.11.0
